refactor: adicionando rota e botão de justificar falta corretamente
